[SYCL][NFC] Cleanup public headers
There are helper function templates which are only used within `source/`
directory, so they were moved in there to reduce the size and complexity
of our public headers.

Some legacy dead code have been removed along the way as well.
