+++ sycl/source/detail/device_impl.hpp
@@ -15,6 +15,7 @@
 #include <sycl/detail/ur.hpp>
 #include <sycl/ext/oneapi/experimental/device_architecture.hpp>
 #include <sycl/ext/oneapi/experimental/forward_progress.hpp>
+#include <sycl/info/info_desc.hpp>
 #include <sycl/kernel_bundle.hpp>
 #include <sycl/platform.hpp>
 
@@ -27,6 +28,40 @@ namespace sycl {
 inline namespace _V1 {
 
 namespace detail {
+
+inline info::partition_property
+ConvertPartitionProperty(const ur_device_partition_t &Partition) {
+  switch (Partition) {
+  case UR_DEVICE_PARTITION_EQUALLY:
+    return info::partition_property::partition_equally;
+  case UR_DEVICE_PARTITION_BY_COUNTS:
+    return info::partition_property::partition_by_counts;
+  case UR_DEVICE_PARTITION_BY_AFFINITY_DOMAIN:
+    return info::partition_property::partition_by_affinity_domain;
+  case UR_DEVICE_PARTITION_BY_CSLICE:
+    return info::partition_property::ext_intel_partition_by_cslice;
+  default:
+    return info::partition_property::no_partition;
+  }
+}
+
+inline info::partition_affinity_domain
+ConvertAffinityDomain(const ur_device_affinity_domain_flags_t Domain) {
+  switch (Domain) {
+  case UR_DEVICE_AFFINITY_DOMAIN_FLAG_NUMA:
+    return info::partition_affinity_domain::numa;
+  case UR_DEVICE_AFFINITY_DOMAIN_FLAG_L1_CACHE:
+    return info::partition_affinity_domain::L1_cache;
+  case UR_DEVICE_AFFINITY_DOMAIN_FLAG_L2_CACHE:
+    return info::partition_affinity_domain::L2_cache;
+  case UR_DEVICE_AFFINITY_DOMAIN_FLAG_L3_CACHE:
+    return info::partition_affinity_domain::L3_cache;
+  case UR_DEVICE_AFFINITY_DOMAIN_FLAG_L4_CACHE:
+    return info::partition_affinity_domain::L4_cache;
+  default:
+    return info::partition_affinity_domain::not_applicable;
+  }
+}
 
 // Note that UR's enums have weird *_FORCE_UINT32 values, we ignore them in the
 // callers. But we also can't write a fully-covered switch without mentioning it
@@ -503,10 +538,6 @@ public:
   /// Queries this SYCL device for information requested by the template
   /// parameter param
   ///
-  /// Specializations of info::param_traits must be defined in accordance
-  /// with the info parameters in Table 4.20 of SYCL Spec to facilitate
-  /// returning the type associated with the param parameter.
-  ///
   /// \return device info of type described in Table 4.20.
 
 #ifdef __INTEL_PREVIEW_BREAKING_CHANGES
@@ -728,7 +759,7 @@ public:
       for (auto &entry : ur_dev_partitions) {
         // OpenCL extensions may have partition_properties that
         // are not yet defined for SYCL (eg. CL_DEVICE_PARTITION_BY_NAMES_INTEL)
-        info::partition_property pp(info::ConvertPartitionProperty(entry));
+        info::partition_property pp(detail::ConvertPartitionProperty(entry));
         switch (pp) {
         case info::partition_property::no_partition:
         case info::partition_property::partition_equally:
@@ -767,7 +798,7 @@ public:
         return info::partition_property::no_partition;
       // The old UR implementation also just checked the first element, is that
       // correct?
-      return info::ConvertPartitionProperty(PartitionProperties[0].type);
+      return detail::ConvertPartitionProperty(PartitionProperties[0].type);
     }
     CASE(info::device::partition_type_affinity_domain) {
       std::vector<ur_device_partition_property_t> PartitionProperties =
@@ -776,7 +807,7 @@ public:
         return info::partition_affinity_domain::not_applicable;
       for (const auto &PartitionProp : PartitionProperties) {
         if (PartitionProp.type == UR_DEVICE_PARTITION_BY_AFFINITY_DOMAIN)
-          return info::ConvertAffinityDomain(
+          return detail::ConvertAffinityDomain(
               PartitionProp.value.affinity_domain);
       }
